+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/CruiseControl.java
@@ -565,15 +565,10 @@ public class CruiseControl extends AbstractModel {
      *
      * @param operatorNamespace                             Namespace where the Strimzi Cluster Operator runs. Null if not configured.
      * @param operatorNamespaceLabels                       Labels of the namespace where the Strimzi Cluster Operator runs. Null if not configured.
-     * @param networkPolicyGenerationEnabled                Activates the generation of Network Policies.
      *
      * @return The network policy.
      */
-    public NetworkPolicy generateNetworkPolicy(String operatorNamespace, Labels operatorNamespaceLabels, boolean networkPolicyGenerationEnabled) {
-        if (!networkPolicyGenerationEnabled) {
-            return null;
-        }
-
+    public NetworkPolicy generateNetworkPolicy(String operatorNamespace, Labels operatorNamespaceLabels) {
         List<NetworkPolicyIngressRule> rules = new ArrayList<>(1);
 
         // CO can access the REST API